modules/ui/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pytest
pytest-asyncio
